+++ packages/material-ui-lab/src/TreeView/TreeViewContext.js
@@ -1,12 +1,3 @@
 import * as React from 'react';
 
-/**
- * @ignore - internal component.
- */
-const TreeViewContext = React.createContext({});
-
-if (process.env.NODE_ENV !== 'production') {
-  TreeViewContext.displayName = 'TreeViewContext';
-}
-
-export default TreeViewContext;
+export default React.createContext({});
